+++ src/iNES/mapper021.cpp
@@ -186,7 +186,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 21;
+uint16 MapperNum = 21;
 } // namespace
 
 const MapperInfo MapperInfo_021 =
