+++ src/iNES/mapper022.cpp
@@ -127,7 +127,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 22;
+uint16 MapperNum = 22;
 } // namespace
 
 const MapperInfo MapperInfo_022 =
