+++ src/iNES/mapper025.cpp
@@ -187,7 +187,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 25;
+uint16 MapperNum = 25;
 } // namespace
 
 const MapperInfo MapperInfo_025 =
