+++ src/iNES/mapper028.cpp
@@ -107,7 +107,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 28;
+uint16 MapperNum = 28;
 } // namespace
 
 const MapperInfo MapperInfo_028 =
