+++ src/iNES/mapper037.cpp
@@ -69,7 +69,7 @@ void	MAPINT	Unload (void)
 	MMC3::Unload();
 }
 
-uint8 MapperNum = 37;
+uint16 MapperNum = 37;
 } // namespace
 
 const MapperInfo MapperInfo_037 =
