+++ src/iNES/mapper042.cpp
@@ -84,7 +84,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 42;
+uint16 MapperNum = 42;
 } // namespace
 
 const MapperInfo MapperInfo_042 =
