+++ src/iNES/mapper047.cpp
@@ -56,7 +56,7 @@ void	MAPINT	Unload (void)
 	MMC3::Unload();
 }
 
-uint8 MapperNum = 47;
+uint16 MapperNum = 47;
 } // namespace
 
 const MapperInfo MapperInfo_047 =
