+++ src/iNES/mapper057.cpp
@@ -57,7 +57,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 57;
+uint16 MapperNum = 57;
 } // namespace
 
 const MapperInfo MapperInfo_057 =
