+++ src/iNES/mapper064.cpp
@@ -186,7 +186,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 64;
+uint16 MapperNum = 64;
 } // namespace
 
 const MapperInfo MapperInfo_064 =
