+++ src/iNES/mapper071.cpp
@@ -61,7 +61,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 71;
+uint16 MapperNum = 71;
 } // namespace
 
 const MapperInfo MapperInfo_071 =
