+++ src/iNES/mapper073.cpp
@@ -105,7 +105,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 73;
+uint16 MapperNum = 73;
 } // namespace
 
 const MapperInfo MapperInfo_073 =
