+++ src/iNES/mapper079.cpp
@@ -51,7 +51,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 79;
+uint16 MapperNum = 79;
 } // namespace
 
 const MapperInfo MapperInfo_079 =
