+++ src/iNES/mapper086.cpp
@@ -68,7 +68,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 86;
+uint16 MapperNum = 86;
 } // namespace
 
 const MapperInfo MapperInfo_086 =
