+++ src/iNES/mapper089.cpp
@@ -34,7 +34,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 89;
+uint16 MapperNum = 89;
 } // namespace
 
 const MapperInfo MapperInfo_089 =
