+++ src/iNES/mapper092.cpp
@@ -57,7 +57,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 92;
+uint16 MapperNum = 92;
 } // namespace
 
 const MapperInfo MapperInfo_092 =
