+++ src/iNES/mapper096.cpp
@@ -61,7 +61,7 @@ void	MAPINT	Unload (void)
 	Latch::Unload();
 }
 
-uint8 MapperNum = 96;
+uint16 MapperNum = 96;
 } // namespace
 
 const MapperInfo MapperInfo_096 =
