+++ src/iNES/mapper101.cpp
@@ -44,7 +44,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 101;
+uint16 MapperNum = 101;
 } // namespace
 
 const MapperInfo MapperInfo_101 =
