+++ src/iNES/mapper118.cpp
@@ -40,7 +40,7 @@ void	MAPINT	Unload (void)
 	MMC3::Unload();
 }
 
-uint8 MapperNum = 118;
+uint16 MapperNum = 118;
 } // namespace
 
 const MapperInfo MapperInfo_118 =
