+++ src/iNES/mapper139.cpp
@@ -87,7 +87,7 @@ void	MAPINT	Reset (RESET_TYPE ResetType)
 	Sync();
 }
 
-uint8 MapperNum = 139;
+uint16 MapperNum = 139;
 } // namespace
 
 const MapperInfo MapperInfo_139 =
